First pass getting some silo tests stable=True.
