+++ src/sentry/conf/server.py
@@ -2852,7 +2852,7 @@ SENTRY_FUNCTIONS_REGION = "us-central1"
 
 # Settings related to SiloMode
 SILO_MODE = os.environ.get("SENTRY_SILO_MODE", None)
-FAIL_ON_UNAVAILABLE_API_CALL = False
+FAIL_ON_UNAVAILABLE_API_CALL = True
 SILO_MODE_UNSTABLE_TESTS = bool(os.environ.get("SENTRY_SILO_MODE_UNSTABLE_TESTS", False))
 
 DISALLOWED_CUSTOMER_DOMAINS = []
