+++ src/sentry/models/userip.py
@@ -3,18 +3,13 @@ from django.core.cache import cache
 from django.db import models
 from django.utils import timezone
 
-from sentry.db.models import (
-    FlexibleForeignKey,
-    Model,
-    control_silo_with_replication_model,
-    sane_repr,
-)
+from sentry.db.models import FlexibleForeignKey, Model, control_silo_only_model, sane_repr
 from sentry.models import User
 from sentry.region_to_control.messages import UserIpEvent
 from sentry.utils.geo import geo_by_addr
 
 
-@control_silo_with_replication_model
+@control_silo_only_model
 class UserIP(Model):
     __include_in_export__ = True
