+++ tests/sentry/api/endpoints/sentry_apps/test_interaction.py
@@ -1,7 +1,7 @@
 from rest_framework import status
 
 from sentry.testutils import APITestCase
-from sentry.testutils.silo import control_silo_test
+from sentry.testutils.silo import region_silo_test
 
 COMPONENT_TYPES = ["stacktrace-link", "issue-link"]
 
@@ -25,7 +25,7 @@ class SentryAppInteractionTest(APITestCase):
         )
 
 
-@control_silo_test
+@region_silo_test(stable=True)
 class SentryAppInteractionAuthTest(SentryAppInteractionTest):
     def test_not_logged_in_not_allowed(self):
         response = self.get_error_response(
@@ -53,6 +53,7 @@ class SentryAppInteractionAuthTest(SentryAppInteractionTest):
         assert response.data["componentInteractions"] == {}
 
 
+@region_silo_test(stable=True)
 class GetSentryAppInteractionTest(SentryAppInteractionTest):
     def setUp(self):
         super().setUp()
@@ -79,6 +80,7 @@ class GetSentryAppInteractionTest(SentryAppInteractionTest):
         )
 
 
+@region_silo_test(stable=True)
 class PostSentryAppInteractionTest(SentryAppInteractionTest):
     method = "post"
