+++ tests/sentry/api/endpoints/test_api_application_details.py
@@ -2,8 +2,10 @@ from django.urls import reverse
 
 from sentry.models import ApiApplication, ApiApplicationStatus, ScheduledDeletion
 from sentry.testutils import APITestCase
+from sentry.testutils.silo import control_silo_test
 
 
+@control_silo_test(stable=True)
 class ApiApplicationDetailsTest(APITestCase):
     def test_simple(self):
         app = ApiApplication.objects.create(owner=self.user, name="a")
@@ -15,6 +17,7 @@ class ApiApplicationDetailsTest(APITestCase):
         assert response.data["id"] == app.client_id
 
 
+@control_silo_test(stable=True)
 class ApiApplicationUpdateTest(APITestCase):
     def test_simple(self):
         app = ApiApplication.objects.create(owner=self.user, name="a")
@@ -29,6 +32,7 @@ class ApiApplicationUpdateTest(APITestCase):
         assert app.name == "foobaz"
 
 
+@control_silo_test(stable=True)
 class ApiApplicationDeleteTest(APITestCase):
     def test_simple(self):
         app = ApiApplication.objects.create(owner=self.user, name="a")
